+++ R/summary.partR2.R
@@ -10,14 +10,6 @@
 #'
 #' @return No return value, prints extended summary of partR2 calculation.
 #'
-#' @references
-#'
-#' Nakagawa, S., & Schielzeth, H. (2013). \emph{A general and simple method for obtaining R2 from
-#' generalized linear mixed-effects models}. Methods in Ecology and Evolution, 4(2), 133-142.
-#'
-#' Newton, R. G., & Spurrell, D. J. (1967).  \emph{A development of multiple regression for the
-#' analysis of routine data. Applied Statistics}. 51-64.
-#'
 #'
 #'
 #' @keywords models